Server/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
numpy==1.26.4
opencv-python==4.8.1.78
Pillow==10.1.0
proto-plus==1.27.0
protobuf==4.25.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
